Fix lcobucci/jwt package dependency version
This package uses the `StrictValidAt` validation constraint which is
only available since the version `4.1` of the jwt package.
Or this version of the jwt package relies on the sodium PHP extension.
If this extension is not installed on the client environment, then the
version `4.0` of the jwt package which, does not provide the requested
class, is installed by default.

Signed-off-by: Eric Villard <dev@eviweb.fr>